[SIG-1006] moved wizard form components and services
