+++ src/signals/incident/components/form/DescriptionWithClassificationInput/style.scss
