+++ src/signals/incident/components/form/Header/style.scss
