+++ src/signals/incident/components/form/MapInput/index.test.js
