+++ src/signals/incident/components/form/PlainText/style.scss
