+++ src/signals/incident/components/form/RedirectButton/style.scss
