+++ src/signals/incident/definitions/kto.js
@@ -3,7 +3,7 @@ import React from 'react';
 
 import { Validators } from 'react-reactive-form';
 
-import FormComponents from '../components/IncidentForm/components';
+import FormComponents from '../components/form';
 
 export default {
   controls: {
