+++ src/signals/incident/definitions/wizard-step-1-beschrijf.js
@@ -2,7 +2,7 @@ import { some } from 'lodash';
 import { Validators } from 'react-reactive-form';
 
 import IncidentNavigation from '../components/IncidentNavigation';
-import FormComponents from '../components/IncidentForm/components/';
+import FormComponents from '../components/form';
 import { checkVisibility } from '../services/format-conditional-form';
 
 export default {
