+++ src/signals/incident/definitions/wizard-step-4-email.js
@@ -1,7 +1,7 @@
 import { Validators } from 'react-reactive-form';
 
 import IncidentNavigation from '../components/IncidentNavigation';
-import FormComponents from '../components/IncidentForm/components/';
+import FormComponents from '../components/form';
 
 export default {
   label: 'Wilt u op de hoogte blijven?',
