+++ src/signals/incident/services/map-dynamic-fields/index.js
